所有功能完成，剩些异常逻辑处理

+++ src/main/java/com/example/lifehealthassistant/service/UserInfoServiceImpl.java
@@ -2,33 +2,51 @@ package com.example.lifehealthassistant.service;
 
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.example.lifehealthassistant.controller.utils.R;
+import com.example.lifehealthassistant.controller.utils.RandomUtil;
+import com.example.lifehealthassistant.dao.UserEmailDao;
 import com.example.lifehealthassistant.dao.UserInfoDao;
+import com.example.lifehealthassistant.domain.Diet;
+import com.example.lifehealthassistant.domain.Disease;
 import com.example.lifehealthassistant.domain.User;
+import com.example.lifehealthassistant.domain.Useremail;
 import org.apache.ibatis.annotations.Update;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
+import java.util.Date;
+import java.util.List;
 
 @Service
 public class UserInfoServiceImpl extends ServiceImpl<UserInfoDao, User> implements UserInfoService{
 
     @Autowired
     UserInfoDao userInfoDao;
+    @Autowired
+    UserEmailDao userEmailDao;
+
+
 
     @Override
     public boolean save(User user) {
         userInfoDao.save_create_diet(user);
         userInfoDao.save_create_disease(user);
+        userInfoDao.save_create_health(user);
+        Useremail useremail=new Useremail();
+        useremail.setId(user.getId());
+        userEmailDao.insert(useremail);
         return (userInfoDao.insert(user)>0);
     }
 
     @Override
     public R saveUserPic(MultipartFile file) {
         if(file.isEmpty()){System.out.println("到saveDietPic循环判断空了");
-            return new R(false,"文件名为空");
+            return new R(false,"文件名为空",null);
         }
         String originFilename =file.getOriginalFilename();
         System.out.println(originFilename);
@@ -44,21 +62,68 @@ public class UserInfoServiceImpl extends ServiceImpl<UserInfoDao, User> implemen
             file.transferTo(dest);
         }catch (Exception e){
             e.printStackTrace();
-            return new R(false,"上传失败，服务器错误");
+            return new R(false,"上传失败，服务器错误",null);
 
         }
-        return new R(true,user);
+        return new R(true,null,user);
     }
 
     @Override
     @Transactional(rollbackFor=Exception.class)
-    public Boolean deleteById(int id) {
+    public Boolean deleteById(String id) {
+        List<Diet> dietList=new DietServiceImpl().getDietAll(id);
+        for(int i=0;i<dietList.size();i++){
+            String filename1=dietList.get(i).getPicture1();
+            if(filename1!=null) {
+                File file1 = new File(filename1);
+                file1.delete();
+            }
+            String filename2=dietList.get(i).getPicture2();
+            if(filename2!=null) {
+                File file2 = new File(filename2);
+                file2.delete();
+            }
+            String filename3=dietList.get(i).getPicture3();
+            if(filename3!=null) {
+                File file3 = new File(filename3);
+                file3.delete();
+            }
+        }
         userInfoDao.delete_diet(id);
+        List<Disease> diseaseList=new DiseaseServiceImpl().getDiseaseAll(id);
+        for(int i=0;i<diseaseList.size();i++){
+            String filename1=diseaseList.get(i).getSympic();
+            if(filename1!=null) {
+                File file1 = new File(filename1);
+                file1.delete();
+            }
+            String filename2=diseaseList.get(i).getMedpic();
+            if(filename2!=null) {
+                File file2 = new File(filename2);
+                file2.delete();
+            }
+
+        }
         userInfoDao.delete_disease(id);
+        userInfoDao.delete_health(id);
         User user=userInfoDao.selectById(id);
         String filename=user.getPhoto();
-        File file=new File(filename);
-        file.delete();
+        if(filename!=null){
+            File file=new File(filename);
+            file.delete();
+        }
         return userInfoDao.deleteById(id)>0;
     }
+
+    @Override
+    public Boolean updateByIdWithoutPs(User user) {
+        return userInfoDao.updateByIdWithoutPs(user);
+    }
+
+    @Override
+    public Boolean updateByIdPs(User user) {
+        return userInfoDao.updateByIdPs(user);
+    }
+
+
 }
